# api/openapi.yaml
openapi: 3.0.0

info:
  title: Curate API contract
  version: 1.0.0
  description: Curate is service that allows you to get crypto currency rates to usd and persist their history

servers:
  - url: http://127.0.0.1:8080

paths:
  /api/v1/rates:
    get:
      description: Returns the latest rates for all available currencies
      operationId: GetAllRates
      tags:
        - rate_ops
      parameters:
        - name: limit
          in: query
          required: false
          example: 10
          schema:
            type: integer
            format: uint32
      responses:
        '200':
          description: Successfully received all rates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rate'
        default:
          description: Rates getting is unsuccessful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/rates/{currency}:
    get:
      description: Returns a rate for specific currency by its currency code (e.g. BTC, ETH, etc.)
      operationId: GetRateByCurrency
      tags:
        - rate_ops
      parameters:
        - name: currency
          in: path
          required: true
          example: "BTC"
          schema:
            pattern: "^[A-Z]{3, 10}$"
            type: string
        - name: limit
          in: query
          required: false
          example: 10
          schema:
            type: integer
            format: uint32
      responses:
        '200':
          description: Rate for specific currency pulled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rate'
        default:
          description: Rate getting is unsuccessful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/v1/health:
    get:
      description: returns 204 if service is up and running
      operationId: HealthCheck
      responses:
        '204':
          description: Service is up and running
        default:
          description: Service is down

  /api/v1/currencies:
    get:
      description: Returns all available currencies
      operationId: GetAllCurrencies
      responses:
        '200':
          description: Successfully received all currencies
          content:
            application/json:
              schema:
                type: array
                format: string

components:
  schemas:
    HistoryPoint:
      type: object
      required:
        - value
        - timestamp
      properties:
        value:
          type: number
          format: double
          example: 12345.6789
        timestamp:
          type: string
          format: date-time

    Rate:
      type: object
      required:
        - currency
        - history
        - value
        - timestamp
        - quote
        - provider
      properties:
        currency:
          type: string
          example: "BTC"
        history:
          type: array
          items:
            $ref: '#/components/schemas/HistoryPoint'
        quote:
          type: string
          example: "USD"
        provider:
          type: string
          format: uri
          example: "https://www.coingecko.com"
        value:
          type: number
          format: double
          example: 12345.6789
        timestamp:
          type: string
          format: date-time
          example: "2021-01-01T00:00:00Z"

    Error:
      type: object
      required:
        - message
      properties:
        message:
          description: HTTP error message
          type: string
          example: "failed to get rates due to..."