# internal/transport/http/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package http

import (
	"fmt"
	"net/url"
	"time"

	"github.com/gofiber/fiber/v2"
	"github.com/oapi-codegen/runtime"
)

// Error defines model for Error.
type Error struct {
	// Message HTTP error message
	Message string `json:"message"`
}

// HistoryPoint defines model for HistoryPoint.
type HistoryPoint struct {
	Timestamp time.Time `json:"timestamp"`
	Value     float64   `json:"value"`
}

// Rate defines model for Rate.
type Rate struct {
	Currency  string         `json:"currency"`
	History   []HistoryPoint `json:"history"`
	Provider  string         `json:"provider"`
	Quote     string         `json:"quote"`
	Timestamp time.Time      `json:"timestamp"`
	Value     float64        `json:"value"`
}

// GetAllRatesParams defines parameters for GetAllRates.
type GetAllRatesParams struct {
	Limit *uint32 `form:"limit,omitempty" json:"limit,omitempty"`
}

// GetRateByCurrencyParams defines parameters for GetRateByCurrency.
type GetRateByCurrencyParams struct {
	Limit *uint32 `form:"limit,omitempty" json:"limit,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /api/v1/currencies)
	GetAllCurrencies(c *fiber.Ctx) error

	// (GET /api/v1/health)
	HealthCheck(c *fiber.Ctx) error

	// (GET /api/v1/rates)
	GetAllRates(c *fiber.Ctx, params GetAllRatesParams) error

	// (GET /api/v1/rates/{currency})
	GetRateByCurrency(c *fiber.Ctx, currency string, params GetRateByCurrencyParams) error
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

type MiddlewareFunc fiber.Handler

// GetAllCurrencies operation middleware
func (siw *ServerInterfaceWrapper) GetAllCurrencies(c *fiber.Ctx) error {

	return siw.Handler.GetAllCurrencies(c)
}

// HealthCheck operation middleware
func (siw *ServerInterfaceWrapper) HealthCheck(c *fiber.Ctx) error {

	return siw.Handler.HealthCheck(c)
}

// GetAllRates operation middleware
func (siw *ServerInterfaceWrapper) GetAllRates(c *fiber.Ctx) error {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetAllRatesParams

	var query url.Values
	query, err = url.ParseQuery(string(c.Request().URI().QueryString()))
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for query string: %w", err).Error())
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", query, &params.Limit)
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter limit: %w", err).Error())
	}

	return siw.Handler.GetAllRates(c, params)
}

// GetRateByCurrency operation middleware
func (siw *ServerInterfaceWrapper) GetRateByCurrency(c *fiber.Ctx) error {

	var err error

	// ------------- Path parameter "currency" -------------
	var currency string

	err = runtime.BindStyledParameterWithOptions("simple", "currency", c.Params("currency"), &currency, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter currency: %w", err).Error())
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetRateByCurrencyParams

	var query url.Values
	query, err = url.ParseQuery(string(c.Request().URI().QueryString()))
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for query string: %w", err).Error())
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", query, &params.Limit)
	if err != nil {
		return fiber.NewError(fiber.StatusBadRequest, fmt.Errorf("Invalid format for parameter limit: %w", err).Error())
	}

	return siw.Handler.GetRateByCurrency(c, currency, params)
}

// FiberServerOptions provides options for the Fiber server.
type FiberServerOptions struct {
	BaseURL     string
	Middlewares []MiddlewareFunc
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router fiber.Router, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, FiberServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router fiber.Router, si ServerInterface, options FiberServerOptions) {
	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	for _, m := range options.Middlewares {
		router.Use(fiber.Handler(m))
	}

	router.Get(options.BaseURL+"/api/v1/currencies", wrapper.GetAllCurrencies)

	router.Get(options.BaseURL+"/api/v1/health", wrapper.HealthCheck)

	router.Get(options.BaseURL+"/api/v1/rates", wrapper.GetAllRates)

	router.Get(options.BaseURL+"/api/v1/rates/:currency", wrapper.GetRateByCurrency)

}